requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Login
Pillow
gunicorn
psycopg2-binary